Fixed compile issues with the BugLife example and now it compiles.
However the code for this is not nice. Some one is going to have to 
inherit this example and clean up the code base and improve the 
documentation on it.

I removed the local arg_parser.h/.cpp that this example was using and
substituted it with the newly added global ArgParser.h/.cpp.  A similar
substitution to use the project-wide ArgParser.h/.cpp needs to happen to 
RescueSim as well (as it has a duplicate version of arg_parser.h/.cpp in it).