Code clean-up and changes to get the user logs to be streamed to a log file. Similar changes need to be carried out to the programmer-log and log views as well.
In the revised code, the Log manages the file streams for logging and reports
changes in starting/stopping of logging via suitable signals.  The views
essentially listen to changes in status and appropriately change the GUI
representation. 

Currently, the views don't listen to the error messages reported by the
log via the errorSavingLog() signal.  The views need to be modified to 
respond to this signal by displaying a suitable error message.