Added a XML marshalling (writing XML) and unmarshalling (reading XML) framework to MUSE-GUI infrastructure (under infa/xml directory). The framework builds on Qt's xmlpatterns library and eases processing XML content in a similar (but simpler) way as JAXB with a few minor restrictions. The detailed documentation in infra/xml/README.txt file for details on the framework and a compelte usage example.
Added a simple Workspace, ServerList, and Server classes to
MUSE-GUI for storing workspace metadata for display.  Currently,
these classes are almost barebones with the ability to read and
write them in XML form using the XML marshalling and
unmarshalling framework.

The MainWindow class has been temporarily modified to use a
default workspace file for testing and operations.  Later on the
default workspace should be replaced with an appropriate initial
dialog that permits the user to choose a workspace.

Also added a top-level README.txt documenting the file
organization paradim used in the MUSE-GUI.