This check-in includes changes to fix marshalling/unmarshalling that was broken after transition to use QObject as base class for XMLElement.
Note that the objects in workspace are not meant to be copied and that
is intentional -- the idea is to prevent duplicate objects from being
created so that the whole GUI works with one consistent set of objects
whose state is updated and maintained correctly.

In order to enable the marshalling/unmarshalling, classes derived from
XMLElement now need to define a helper class using the newly
introduced DECLARE_METATYPE_HELPER macro. Then the classes can be
registered via call to Qt's qRegisterMetaType() as done in
Workspace::registerClasses() method.

See changes towards the bottom of Server.h and Server.cpp for how to
use the macros.  The documentation in XMLElement.h has also been
updated to reflect the use of DECLARE_METATYPE_HELPER macro.