just a couple changes to button behaviour on the workspace dialog, changed slot from released() to clicked(), not very important
There was a problem with not being able to create a new workspace if you chose a file
with the file browser in the WorkspaceDialog, a Q_ASSERT in the method getType() in the
XMLElementInfo class would fail when the ServerList object was initialized in the Workspace constructor.  
I believe the problem is because ServerList stores a QList of type XMLElemnt*.  
To fix this issue I made two changes:

1) registered the class XMLElement in the registerClasses() method in Workspace
2) called registerClasses() in the createWorkspace() method so that all of the required
classes are registered before the Workspace constructor is called

These two changes seem to have fixed the issue because creating a new workspace now works
correctly, and you are able to add servers with the ServerWizard that will persist through
restarts of the gui.