Added a comletely new visualization for SEARUMS simulation results. Circos (http://circos.ca/) illustrating the summer and wintering inter-country (note that intra-country is not shown as it is not relevant for this study) pathways and influential clades.
The circos look good. However, the sizes of the images are much bigger
than anticipated (I was anticipating about 1-2 MB but got 11 MB file
size).  This could be due to the number of links in the data. Possibly
the links can be converted to ribbons
(http://circos.ca/documentation/tutorials/links/ribbons/lesson) to
reduce the file sizes.