Added more results to the abstract and author summary sections.
Revised title to better refelect the research discussed in the paper.

Replaced the methods figure with a recent version. The source files
for updated PDF are missing from the archive and Neil needs to check
them in.

The paper needs one more round of critical reading (can't be me as my
eyes will not catch them) and after that it will be ready for submission.