Included a more automatic *but not fool proof* way of generating final figures to be submitted. The approach is:
1. The top-level figure captions were changed to include file names as
in: \caption{Blah} was changed to \caption{pic.pdf}{Blah}.

2. These files names are shown only when they are turned on by command-line arguments to `pdflatex` command.

3. A extract_figures.sh bash shell script uses this command-line
option to latex the article, extracts captions in the form "Figure
1. (indon.pdf, egy.pdf, viet.pdf) ..." and uses that information to
convert the referenced PDF files to TIFF files in
figures/final_figures directory with file names appropriately renamed.

The issue is that the command relies on PDF to text conversion which
is currently not fool-proof as the formatting in text can be different
than the PDF. For example, "Figure 17" does not convert correctly and
needs to be manually converted.