Modified Antigen to use the species-aware host populations after some testing.  There were a couple of bugs that I discovered during testing and fixed it. For testing I used the repeatable random number set. The result from the simulations with and without species looked very close and consequently I "think" my implementation is good.
NOTE: The new implementation does not grow the population outside the
brooding season like the earlier code was doing. This causes some
differences in the result.

Nevertheless, if you get a chance to compare/contrast the earlier (in
HostPopulation.java) and new (in HostSpeciesPopulation.java) versions
as a code-review or test that would be wonderful.

In addition to using the new species-aware host populations, I
reworked and streamlined the command-line argument processing code in
Antigen.java.  I added the following two new parameters:

1. paramFile: Override the default parameter file path.
2. outputDir: Override the default output directory.

The new HostSpeciesPopulation.java honors the above two
parameters. However, the earlier HostPopulation.java does not use
these 2 commnad-line arguments.