One of the sources of memory errors in Haplos was schedule generation. In some cases, the schedule generation exceeds week time frame of 1008 steps (10 minute steps for 24 * 7 hours). This causes schedule generation to have issues.
Unfortunately, the schedule generation method is an unwieldy 1500
lines of code.  So for now, as a work around, I added a check in
Location.cpp to check and handle `endTime > 1008` to plug the memory
error.

Note that this is a workaround.  The correct solution is to clean up
schedule generation to ensure that the end time is never > 1008.  I
will create a ticket for this clean-up.