Streamlined the testing of scheduler queues in muse. This revised code has a new functional test in MuseDatStructure/main.cpp.  The test cases uses the scheduler to create the actual queue being tested. It adds several agents and events to the queue (via scheduler API) and prints the events to verify that the events have been added to the queue and are stored in the expected order in the different layers of the scheduler queue.
NOTE: A key change in this check-in is the addition of the 'prettyPrint()'
method to the scheduler's public API.  This method merely calls the
corresponding method on the scheduler queue.

The netbeans project assocaited with the tester is not checked-in to avoid
issues with cross platform portability and interoperabilty warnings/errors
generated by NetBeans.

Note that the 'MuseDatSturecutre' directory is incorrectly named. Maybe it
will be changed to 'QueueTests' or something along those lines in a future
commit.