Added a EVENT_RECYCLING option to the kernel --
Profiling the code with different queuing data structures showed that
a large number of memory allocations/deallocations were arising from
event creation and deletion.  These calls accounted for at least 20%
of the simulations runtime.

In order to reduce this overhead (cannot be completely eliminated) --
a reasonably generic event recycling option has been added to the
kernel.  Currently it works only for muse::Event (although it can be
extended to other types of events) in the following manner:

  - When an event's reference count drops to 0, the event is added to
    a recycle store based on its size.

  - Whenever an event is created, the recyle-store is checked to see
    if it already has an event of the given size.  If so, the event is
    used.

The event recycling option reduces runtime of fine grained (that is
time taken by model to process an event is very low) simulations,
specifically PHOLD with granularity of 0, by about 20%.