OK, I think computer science and big-O makes sense :-)
Added additional statistic indicating the number of compares that are
estimated to be performed by ladderQ and heap2tQ.  The number of
compares for the two queues is estimated at the end of a run using the
following formulas:

fixHeap_compares     = avg_fixHeap_compares * num_samples
compares for heap2tQ = total_schedules * log2(#agents) + fixHeap_compares

# LadderQ uses a heap for sorting the bottom (n*log(n)). So:
compares for ladderQ = log2(avg_bottom_size) * events_in_bottom

Difference between the above number of compares for the two queues
seems to provide a good indicator of which queue is going to be
faster.

Still need to build a better understanding/mapping between PHOLD
command-line parameters and the "number of compares" reported by the
simulation(s) in different configurations.