Adding batch of sequential simulation statistics for PHOLD 1000 agents. Unfortunately, the simEndTime of 500 is a bit short for this model and some of the numbers < 1 second which is not good for a paper.
So this set needs to be repeated with a simEndTime of 5000 to have a timing
that is > 1 sec. 

The data for 1 million agents has a lot of variability in it on Red Hawk 
with some jobs running in ~100 seconds while exactly the same job exceeds
 1 hour. I am not sure if larger memory footprint is causing some issues.
So for the paper I am not going to pursue the 1 million agent model.  
Possibly the 1 million agent data is something that can go with the 
Thesis data.