This check-in fixes a major bug/issue with the way the Bottom of ladder queue was managed -- with good and bad news --
   Prior to this check-in, the bottom was managed based on the
   assumption of a linked list where sorting was done by
   searching+insert -- this is a linear scan of bottom which is slow,
   particularly when bottom is large, which is the case when there are
   a lot of concurrent events.

   This revision replaces linear with binary-search+insert and
   directly using std::sort for VectorBucket, which is based on
   std::vector

With this change the LadderQ performance almost doubles and the
performance of heap2tQ is realized only when there are about 10 or
more concurrent events in sequential simulation.

However, the LadderQueue still has performance issues in parallel
where the heap2tQ recovers from rollbacks in a more robust manner than
the ladderQ.  Possibly, a tight scheduler time window could address
some (but not all) of the rollback issues with LadderQueue.