Checking-in the implementation for the top-level 2-tier ladder queue. The implementation is pretty close to the ladder queue class with necessary modifications to use 2-tier structures consistently throughout.
With this check-in the 2-tier ladder queue seems to work fine with
sequential simulations.

However, it fails an assertion (and aborts) in parallel. Have to
troubleshoot that issue before being able to compare its performance.

This implementation still *does not* have a potential dirty-bit/time
optimization to short circuit repeated cancellation for the same
sender ID.  Only after the base 2-tier queue works in parallel, the
dirt-bit/time optimization can then be added.