Implemented the recent set of changes to ladder queue to 2-tier ladder queue.  The changes include:
1. Clean-up of empty ladder rungs was moved out of `recurseRung`
   method into `populateBottom` method. These changes make the rung
   clean-up closer to the paper description and also ensure that the
   ladder is as short as possible.

2. Added functionality to move bottom back into the ladder.  This was
   missing earlier and this functionality is needed in one edge case
   that happens.

3. Added methods and counters to detect when bottom is moved into
   ladder.

Still the two versions struggle in scenario where there are are a lot
of rollbacks near the epoch -- that is most of the scans happen in the
bottom.  The observation suggests that we may need to reorganize
bottom into 2-tiers.  Have to think about that design a bit before
implementing it.