One of the issues that I have been trying to troubleshoot for the last 24 hours (or so) is that in certain scenarios the ladderQ and 2-tier ladderQ would start running very-very-slow.
In an effort to troubleshoot this issue, I started with trying to
reduce the sequential runtime of 2-tier ladder queue which is about
10% to 15% slower than the default ladder queue implementation.  One
thought that I ad was that it could be attributed to the memory
allocation overhead for the extra sub-buckets.

So I have modified 2-tier ladder queue to reuse entries in the ladder
rather than recreating them.  The 2-tier ladder queue now uses a nRung
variable to track the rungs used in the ladder. The rungs are never
popped-off -- instead when rungs are added nRung is incremented and
when rungs are "logically removed" nRung is decremented.

This change (i.e., reusing rungs) did not make a big difference in the
sequential runtime of the 2-tier ladder queue.  However, interestingly
it made general runtime of parallel simulations a lot more reliable,
particularly in cases where there are many (i.e., over 500)
bottom-to-rung operations.  So it seems like a key issue plaguing the
ladder queue implementations could have been memory allocation
overhead -- but that was not obvious from profile data for some reason.

With this change the 2-tier ladder queue is performing fine -- i.e.,
without major hiccups noticed earlier and operates in a smooth manner
similar to heap2tQ.

Next, I need to migrate this change (i.e., reuse rungs) to regular
ladder queue (maybe sometime tomorrow) to ensure fair/consistent
comparisons.