In addition to comparing runtime, I also added charts to compare peak memory footprint.  This chart intentionally uses 2 different y-axis (actually they are 2 different charts placed next to each other).
The memory chart is interesting.  Looks like the linked-list has a ~6x
smaller peak memory footprint when compared to vectors.  This is
because vectors double their size when they grow. So when we have lots
of buckets each with 2x vectors that shows up as a ~6x increase in
memory footprint.  However, the extra memory would get amortized as
the number of events grow. Nevertheless, an interesting chart, given
that the configuration that is slowest is the one that is most memory
efficient one.

I tweaked the size of the timing comparison chart to be the same as
the memory footprint chart, to ensure things looks nice in case we put
both charts next to each other in a paper.