When analyzing the sequential simulation results from different PHOLD models it is clear that the maximum number of rungs in the ladder plays a role in the effective performance of the queue.  Too few rungs seems to result in inserts into bottom which slows the simulation down considerably.  On the other hand, if the rungs are too many then cost of moving events from rung-to-rung seems to be start dominating.
So the maximum number of rungs is a balance and I modified the two
ladder queues to permit MaxRungs to be set via --lq-max-rungs (same
for both) command-line argument.

In addition, another thought that comes to mid is to possibly avoid
splitting buckets if the bucket width is smaller than a given
threshold.  That check was in the ladder queue implementation earlier.
However, I removed that check as it did not perform well in sequential
or parallel simulation. So some more analysis would be needed on the
long run to figure out an optimal setting/threshold for: MaxRungs
and/or MinBucketWidth.