Plots of some of the key statistics collected via Linux perf from *sequential* runs of ph3 and ph4 PHOLD configurations.  The charts focus on caching properties of the queues, specifically: cache_misses per event and cache_misses per instruction.
In addition, a chart of instructions per event is also plotted.

Overall the charts look fine. However, one set of data points that
seem a bit off is instructions/event between ladderQ and 2tLadderQ

Ideally, in sequential simulation, the number of instructions for
2tLadderQ should very close but slightly higher than ladderQ (due to
extra sub-bucket operations) like it is for ph3_instrs_per_event_delay_1.pdf

However, for others it is counter -- the 2tLadderQ seems to be
reporting slightly fewer instructions than ladderQ which is counter
intuitive. Maybe the sampling rate for perf needs to be adjusted to
get better/accurate instruction counts based on the results from
callgrind below:

Running with callgrind seems to provide the correct information --
that is ladderQ does indeed run fewer instructions as summarized below:

$ valgrind --tool=callgrind ./phold --time-window 10 --rows 100 --cols 100 --scheduler-queue ladderQ --simEndTime 500 --delay 10 --selfEvents 0.250 --eventsPerAgent 20
ladderQ   : I   refs:      166,474,377,371

$ valgrind --tool=callgrind ./phold --time-window 10 --rows 100 --cols 100 --scheduler-queue 2tLadderQ --2t-ladderQ-t2k 1 --simEndTime 500 --delay 10 --selfEvents 0.250 --eventsPerAgent 20
2tLadderQ : I   refs:      167,565,574,157