Created an initial repository for the DHS Foresight project. The SVN repository will contains all the of the artifacts (and their evolution).  The archive can be checked out using your Miami University ID (MUid) and credentials via:
$ svn checkout --username MUid https://pc2lab.csi.miamioh.edu/svn/pc2lab/foresight foresight

The archive can also be viewed via the Trac project site at:
https://peace.csi.muohio.edu/trac/pc2lab

NOTE: The commit-logs will be used to generate reports. Consequently,
it is important to frequently commit to SVN -- possibly at the end of
4 hours of work, immaterial of whether the functionality is
operational.

Another way to look at it -- No SVN commits implies no work was done
and consequently no salary.  So SVN commits are a critical aspect of
working, particularly as a team.