Added code to upload time-based snapshots of a reference model to the database. The model generation uses an external C++ binary in /usr/loca/bin/sbfftool on pc2lab-phi server to try to speed up the JSON generation. The C++ binary processes a large binary model (~1.6 GB) and generates JSON. The PHP side parses the text output and loads chunks of JSON into the data base.
The PHP code is run as:

$ time php artisan -vv  model:load --name="chikv" --desc="Reference Chikungunya model" --user=1 ~/temp/full_video_model.sbff

Currently, the C++ side takes < 3 minutes to generate the JSON from the 
~1.6 GB binary. Unfortunately, the PHP side takes > 42 minutes to load the
data into the database.

So during this time the model is not in a consistent state in the database.
I think it would be best to have a 'available' boolean field in the 'models'
table to temproarily mark the model as being unavailable while the data 
is being loaded. Once the data is loaded into the db this flag can be set
to indicate that the model is ready to be served out.