Added a one-to-many relationship between model and snapshots & entities tables
Added a simple route to get the metadata for a model. 

However, the Laravel framework encodes the metadata thinking it is a string.
This casues the metadata to be encoded and returned. Instead it is preferred
that the metadata is returned without encoding. Not sure how to get Laravel
to not encode metdata.