More updates to conversion. Updated Agent to include all data members, EventQueue now uses Event instead of AgentEvent, ThreeTierHeapEvent Queue is implemented up to managing heap with fibHeapPtr which I still need to figure out how to implement with java.
I also came across an issue with enqueueEvent as PriorityQueues in java are not sorted and thus can't be searched in log time. Original implementation searches for a matching bucket to put the event in if possible, such a search with java implemetnation has to be linear time. May need to use a smarter data structure than PriorityQueue