Fixed null pointer exceptions due to missing initialization of instance variables with 'new' objects (of the same type) in various points in the code.
Commented out several debug output statements.

Now the benchmark seems to be running for very trivial cases producing
the same output as the C++ side. However, more testing with different
configurations is needed to determine if the Java version of
LadderQueue is operating per-design.

However, this check-in at least gets something working on the Java
side now.