Update ThreeTierHeapEventQueue to utilize a FibonacciHeap to hold agents which allows log(l) time to update heap on an enqueue or dequeue event. (l number of LP's)
Still have an issue with the EnqueueEvent method as I can't seem to find a good work around to C++'s heap based implementation where the elements were actually sorted and could thus be searched in log time to find matches for concurrent events. As a result, the search through the tier2 container to find the container to place a new event is still done in linear time which is not optimal.