Set up the framework for handling roles with users. "Manage People" page now shows a list of users and will soon include forms for managing roles. (database is ready for roles but need to add the forms to actually control them from the UI).
A level deeper would be to include a "actions" table that lets you manage what the different roles can do, this is a low priority however.

Also added a "admin" flag to users as admin will be consider separate from the roles and will have full access to everything. The admin flag is set to false by default, but currently has no function.