Added a toy epidemic model for further enhancement and use for testing CPU/GPU implementations. Currently the model does a basic sequential simulation type setup. There is some capability to save state to test rollback behaviors; but currently that code-path is untested.
Moreover the toy model currently lacks the actual SEIR ODE/SSA code path
and that needs to be added.

Lastly, the configurability of this simulation is very low with several
constants coded into the source.  Eventually these would need to be pulled
out to a suitable configuration file that can be used to initialize the
model and its epidemic parameters.