Added code to print the actual number of buckets finally left in the ladder. This is an indication of the overall number of buckets used.
With large timestamps (such as those seen in waterfowl-migration model) some
of the rungs in the top become quiet large with over 5000 buckets in cases
with rollbacks. 

Maybe this could be addressed but trying to ensure large bucket widths; but
I am not going to do that yet because that would be deviating from the 
original Ladder queue algorithm by Tang et. al.  Instead I will do it 
2tLadderQ to see how it works out.