On OSC clusters, with mvapich, the cost of calling MPI_IProbe to check for events is rather expensive (more expensive than open-mpi's implementation).  Consequently, I have added a linear-backoff (rather than an exponential-backoff) approach to check for incoming MPI messages.
The exponential-backoff was generating too coarse a threshold and was
not handling scenarios with rollbacks very well.  On the other hand,
linear-backoff (increments of 5, the 5 is some magic number) seems a
bit more stable under initial debug/testing.  However, more testing is
needed to see if this linear-backoff is going to operate well under a
broad range of conditions.