This check-in adds functionality to redistribute unused NUMA-memory chunks between threads when imbalance is detected in NUMA-memory recycler.
Problem: When NUMA-aware event recycling is used in multi-threaded
simulations, the size of unused event pool can be skewed -- that is,
some threads end up having a large fraction of events that are not
used.  Furthermore, the "NUMA Recycler %hits" statistic would degrade
well below 90% to 70%.  This issue is most obvious when shared events
are not used, but can occurr in even with shared events depending on
the communication patterns.</p>

Solution: Now, the NumaMemoryManager has the ability to redistribute
unused memory chunks, evenly to different threads.  This message is
used to send unused chunks to different threads using the normal
inter-thread communication mechanism. At the end of each garbage
collection cycle, the multi-threaded simulator calls the redistribute
method in the NumaMemory manager (note that this happens on each and
every thread).  The NumMemory manager checks to see if it has more
recycled chunks than the NUMA-blocks it actually allocated.  If so, it
redistributes the unused memory chunks evenly to all the threads using
this message.

With this change the NUMA memory usage and recycler-hits seems much
better, typically over 95% recycler-hits.