Fixed a long lurking bug (took a long time to identify the source of the issue as well) in the core kernel with handling anti-messages -- prior to this commit anti-messages were not 100% correctly processed in the kernel -- specifically, anti-messages with sentTime before LVT but recvTime after LVT did not trigger rollback processing, like it should have.
Now rollback processing for anti-messages is correctly triggered based
on sentTime of anti-messages in Scheduler.cpp.  This change required
changes to how anti-messages were sent when cleaning-up output queues
-- now the lowest-sentTime and lowest-receiveTime for each receiver
are tracked and used when sending anti-messages to cancel output
events.

Tracking the bug required updates and fixes to some of the
debug-code-paths in some of the queues as well.

The debugging experience clearly indicates that we need models that
can be used as functional testing (most of our models are for
performance testing) and we need to design some reliable functional
testing models (say a digital-logic circuit model).