Cleaned up routes and made a mock up of a very simple dashboard for models and analyses.
Site now has a public facing landing page and the default page once logged in is the mocked up dashboard. Work still needs to be done to make the landing page look pretty, and the dashboard is not functional at the moment but should be fairly easy to wire up with javascript.