Replaced the concept of "Groups" with the concept of "Roles" and implemented the spatie/laravel-permission package.
This package provides the concept of "roles" and "permissions" and comes with many important features out of the box. Permissions can be granted to a role or directly to a user. This provides flexability as "permissions" will eventually represent the various permissions on a model or analysis, thus allowing us to grant levels of access to various entities on both a group/role level and also at the individual level.

For example, an analysis may have admin, edit, and view permissions. The creator of it will have admin permissions directly by default, and can then delegate other permissions to both groups (roles) or users as they see fit. Checking permission would be as simple as $user->can('edit_analysis_12);

Creating roles and adding users to those roles is already implemented and working under the "Manage Users" page.

You will need to do a "php composer.phar update" and a "php artisan migrate" to get this working on existing instances.