muse compiles, builds, and runs (sequentially) with mpi-mt-shm integrated
Hit the biggest obstacle on memory management, and numa based event recycling is not fully integrated. Since this version of multi threading is shared memory, much of the numa logic may not be needed, but will need to work with Dr. Rao to decide on best approach.

PHOLD hit a seg fault when multiple threads were used, more work still needs to be done before muse can fully operate as desired for mt-mpi-shm.