MPI-MT-SHM runs multi threaded to completion and is stable.
This version is fully thread safe, however it requires use of --scheduler-queue badMTQ which fully locks on all critical operations. Next step is to make this structure thread safe as well.

Upon testing, the default simulation was unaffected but mpi-mt seems to no longer work. A good amount of time was spent debugging this, but I can't seem to isolate what caused mpi-mt to stop working and it doesn't seem to be related to this specific commit. Due to time constraints checking this in anyway. 
To run, the flags --simulator mpi-mt-shm and --scheduler-queue badMTQ must be present.