Added a significantly recoded chloropleth map visualization code.
Currently this is used by analysis setup that can be viewed via 
/analyses/setup/1 URI.

The Map has most of the features of the original one except the following --

The speed of setting colors for entities is extremely slow -- practically
unusable.  This issue needs to be investigated to make it fast.

Popups are currently not implemented.