Finally got to work on the input form -- my plans got setback by a few days as I had to rework the map/model rendering code from scratch.
Added functionality to setup parameters and information to be used for
analysis.  Now the form is almost complete and is able to generate a 
valid submission. However, the backend submission processing logic is not yet
in place to actually start and do the analysis.

One big feature that is missing is selection of entities for analysis by
simply selecting entities on the map. This feature can be operated on 
independently by some other team member while I focus on backend stuff.