Checking in most recent version of LockFreePQ for code review.
There is a very minimal benchmark that is built into the queue class and can by run using the compile command in the top of LockFreePQ.h and then running the executable with an argument for the number of threads to spin off.

I've been working to integrate this queue into a new ThreeTierSkipMTQueue.cpp (3tSkipMT) but that won't be ready for review until later today or tomorrow.