Templatized LockFreePQ to use any arbitrary Key, Value, and Comparator type. The queue is now fully generic
Preliminary benchmarks for this queue are great. When compiled with O3 on toy benchmark I'm getting significant speed up with up to 4 threads compared to sequential.

Garbage collection is the big issue. While the benchmarks have gotten faster with refactoring, the number of seg faults has increased when the queue is run with a large number of values. That being said, the toy benchmark is worst case scenario as it is essentially continuous contention on the structure over millions of values, but queue should slow down before seg faulting every time.

I've begun to look into memory ordering on atomic structures as a possible cause of the issue, however the default for atomics is already the most restrictive mode of operation. It's still possible I'm making assumptions about memory ordering for atomics that the library doesn't actually provide, and further work is necessary to make the epoch based garbage collection reliable in the large scale.