PHOLD runs to completion with simulation mt-mpi-shm and scheduler-queue 3tSkipMT. The real Agent class was made naively thread safe with locks, and garbage collection does not work fully as the finalize methods throw asserts due to event reference counters not being fully decremented. The simulation is also heavily bogged down by locks and contention, removing most all of the benefits of a thread safe data structure.
This check in will be rolled back to revert core muse files to avoid clutter, as the integration effort for multi threaded PES scheduler queue has been deemed out of scope for the multi-threading effort. This check in will serve purely to demonstrate a snapshot of muse with "most" of the multi threaded integration finished and a version of muse that at least runs multi threaded.

The queue robustness benchmark simulates PHOLD without the additional complexities of Agent state saving, GVT estimation, and non-simple garbage collection. While muse has a lot of work done to allow for multi threaded operation, that part of the project will be future work.

Lastly, this code doesn't compile without openCL header files installed, to avoid conflicting with Harrison's work. Those parts of the code need to be commented out at this version to compile.