Fully revised draft of thesis in preparation for review and feedback from Dr. Rao.
This version has been edited from the ground up, though I am still waiting on some extra data for the results section. Specifically, I am collecting a large amount of data to convert granularity factor into a reliable amount of elapsed CPU time, and I am also re-doing the comparison of # LPs to runtime as the initial scalability results used a granularity factor of 0, which has since been found to be unrealistic and not useful for measuring multi threaded performance. The hope is to use a granularity factor that is represenative of real world scenarios.

Conclusion and results are both not very refined (the rest of the paper is), so big picture feedback on both of those sections would be appreciated so I can add more detail per your review.

Thank you in advance for the time spent on this.
~Matt