Fully revised draft from previous revisions and also added a section for data collected from the USE simulator.
The USE data doesn't look very good, but it's hard to tell if it is their simulation that is causing it, the changes I made, the environment we're running it in, or something else. I'm guessing it would look bad to publish someone elses work in a bad light if it isn't 100% accurate, and there are a lot of variables at play with the issues in their simulators ability to acheive speedup. I'll let you make the call on how to proceed with that section.

I would recommend just looking at the change log to review this edit, as many small changes were made throughout the paper and there's no need to re-read the whole paper.

I will be reading the paper cover to cover right now to catch any small issues. Again, I am available in person to defend the week of July 9-13. If that time range won't work please let me know asap as I have a flight leaving Oxford that weekend.

I can't apologize enough for the delay on this. Hopefully anything that is left is small, but if the timeline doesn't work out I take full responsibility for it, and please don't go out of your way to rush the process.