Small revisions and addition of USE comparative anslys in results section.
This check in is the version that was sent to committee, except for the small change to Figure 6.5 label from email this morning which is also included.

There may be some more small edits to go as I read the paper for a final time, but for the most part this commit represents the final version of the paper.