Initial integration with BLTI and oAuth check is done and seems to be working fine.
Now, making changes to accommodate different base URLs -- this is needed
to enable/separate production vs. testing code paths without having to
duplicate servers etc.  The basic operations seem to be working but need to
rebuild the docker image in order to accommodate the change in base URL
in config.php. Maybe this base url should become an environment variable?