Added functionality to display influence metrics for model parameters. The influence metrics enable identification of influential factors that can be controlled to mitigate progression of epidemics in different countries.
The Popup obtains the data to be displayed via an AJAX call and displays it
in a table.

Problem: Currently, the Popup UI is broken -- it is mishmash of bootstrap 
css and poorly setup css in popup_panel.css.  This mishmash needs to be 
cleaned up to use purely Bootstrap UI with minimal homegrown css.