Added feature to remap population from one given population ring to another to adjust parts of the model where minor alignment differences between OSM data and LandScan grids occurs.
The remapping has been manually checked and validated.

The remapping feature has been implemented by extending the building
ignoring code. So this code path was renamed to "model adjustments"
since it now includes both building ignoring and population remapping
features.  This change in nomenclature is reflected in the
input files as well.