This commit has an important API change to address many warnings related to the use of dynamic exceptions, which have been deprecated in C++11 (and removed in C++17).  This check-in changes the core MUSE kernel and the examples -- "initialize() throw(std::exception)" is now changed to "initialize()" -- thereby removing the dynamic exception specification.  Note that the method can still throw exceptions but it is implicitly assumed rather than explicitly being stated.
The above change required changes to all of the associated methods in
the kernel to avoid many warnings. On one hand side, the change is
somewhat small (just one liner changes in both headers and source) but
it touches a lot of files and consequently, the change set is large.